생애 최초 JPA 사용하기 : 26강. By 앞에 들어갈 수 있는 구절 정리 - exists

+++ src/main/java/com/group/libraryapp/service/user/UserServiceV2.java
@@ -44,9 +44,15 @@ public class UserServiceV2 {
 
     public void deleteUser(String name) {
         // SELECT * FROM user WHERE name = ?
-        User user = userRepository.findByName(name)
-                .orElseThrow(IllegalArgumentException::new);
+//        User user = userRepository.findByName(name)
+//                .orElseThrow(IllegalArgumentException::new);
+//        userRepository.delete(user);
 
+        if (!userRepository.existsByName(name)) {
+            throw new IllegalArgumentException();
+        }
+
+        User user = userRepository.findByName(name);
         userRepository.delete(user);
     }
 }
